implement Serializable in beans

+++ src/main/java/com/univ/dao/Identificator.java
@@ -3,5 +3,5 @@ package com.univ.dao;
 import java.io.Serializable;
 
 public interface Identificator<PK extends Serializable> {
-    int getId();
+    long getId();
 }
